Add unit tests for term-level queries in dynamic query DSL
- Implement ExistsQueryTest to validate the creation of exists queries within must clauses, including handling of null fields and boost settings.
- Implement RangeQueryTest to ensure range queries are correctly generated in must clauses, with checks for boost and query names.
- Implement TermQueryTest to verify term queries in must, filter, mustNot, and should clauses, including handling of null or empty values.
- Implement TermsQueryTest to validate terms queries in must, filter, mustNot, and should clauses, ensuring proper handling of null or empty values and boost settings.
