+++ src/main/kotlin/com/github/silbaram/elasticsearch/dynamic_query_dsl/queries/termlevel/TermQuery.kt
@@ -1,0 +1,24 @@
+package com.github.silbaram.elasticsearch.dynamic_query_dsl.queries.termlevel
+
+import co.elastic.clients.elasticsearch._types.query_dsl.Query
+import co.elastic.clients.elasticsearch._types.query_dsl.TermQuery
+
+fun termQuery(
+    field: String,
+    value: String?,
+    boost: Float? = null,
+    _name: String? = null
+): Query? {
+    return if (value.isNullOrEmpty()) {
+        null
+    } else {
+        val builder = TermQuery.Builder()
+            .field(field)
+            .value(value)
+
+        boost?.let { builder.boost(it) }
+        _name?.let { builder.queryName(it) }
+
+        builder.build()._toQuery()
+    }
+}
